Almost buildable state
