+++ src/CMakeLists.txt
@@ -1,1 +1,1 @@
-/opt/ros/lunar/share/catkin/cmake/toplevel.cmake
+/opt/ros/kinetic/share/catkin/cmake/toplevel.cmake
