+++ src/jhonny5/jhonny5_main.h
@@ -1,7 +1,10 @@
 #include "ros/ros.h"
 #include "std_msgs/Bool.h" 
 #include "std_msgs/String.h" 
+#include <geometry_msgs/Twist.h>
 #include <actionlib/server/simple_action_server.h>
+#include <laser_geometry/laser_geometry.h>
+#include <sensor_msgs/LaserScan.h>
 #include <jhonny5/MoveOrder.h>
 #include <sstream>
