+++ src/jhonny5/package.xml
@@ -40,6 +40,9 @@
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>actionlib_msgs</build_depend>
+  <run_depend>actionlib_msgs</run_depend>  
+
 
 
   <!-- The export tag contains other, unspecified, tags -->
